# src/lib/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduWeb API",
        "description": "Documentación de la API para la plataforma educativa EduWeb, diseñada para ofrecer una experiencia de aprendizaje en línea integral.",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/users/{id}/courses/{courseId}/exams/{examId}": {
            "get": {
                "tags": ["Cursos"],
                "summary": "Obtener detalles del examen",
                "description": "Recuperar detalles de un examen específico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "user_exam_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "user_course_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "exam_state": {
                                                    "type": "string",
                                                    "enum": ["progress", "completed"]
                                                },
                                                "userAnswers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "user_exam_id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "answer_id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "answer": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "answer_id": {
                                                                        "type": "string",
                                                                        "format": "uuid"
                                                                    },
                                                                    "question_id": {
                                                                        "type": "string",
                                                                        "format": "uuid"
                                                                    },
                                                                    "answer_text": {
                                                                        "type": "string"
                                                                    },
                                                                    "is_correct": {
                                                                        "type": "boolean"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del examen obtenidos exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener los detalles del examen"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Cursos"],
                "summary": "Enviar respuestas del examen",
                "description": "Enviar respuestas para un examen, verifica si el usuario ha aprobado el examen y actualiza el estado del curso",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuestas enviadas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Respuestas del examen procesadas exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al procesar las respuestas del examen"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/courses/{id}": {
            "get": {
                "tags": ["Cursos"],
                "summary": "Obtener curso por ID",
                "description": "Recuperar detalles de un curso específico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Course"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Curso obtenido exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener el curso"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_name": {
                        "type": "string"
                    },
                    "user_lastname": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "user_image": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                }
            },
...
            "Role": {
                "type": "object",
                "properties": {
                    "role_id": {
                        "type": "integer"
                    },
                    "role_name": {
                        "type": "string"
                    }
                }
            },
...
            "Course": {
                "type": "object",
                "properties": {
                    "course_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "teacher_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "course_name": {
                        "type": "string"
                    },
                    "course_description": {
                        "type": "string"
                    },
                    "course_image": {
                        "type": "string"
                    },
                    "course_difficulty": {
                        "type": "string",
                        "enum": ["básico", "intermedio", "avanzado"]
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "teacher": {
                        "$ref": "#/components/schemas/User"
                    },
                    "exam": {
                        "$ref": "#/components/schemas/Exam"
                    },
                    "blocks": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "block_id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "course_id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "block_title": {
                                    "type": "string"
                                },
                                "block_description": {
                                    "type": "string"
                                },
                                "block_order": {
                                    "type": "integer"
                                },
                                "lessons": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "lesson_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "block_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "lesson_title": {
                                                "type": "string"
                                            },
                                            "lesson_video": {
                                                "type": "string"
                                            },
                                            "lesson_description": {
                                                "type": "string"
                                            },
                                            "lesson_duration": {
                                                "type": "string"
                                            },
                                            "lesson_image": {
                                                "type": "string"
                                            },
                                            "lesson_order": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "Exam": {
                "type": "object",
                "properties": {
                    "exam_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "course_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "exam_title": {
                        "type": "string"
                    },
                    "exam_description": {
                        "type": "string"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "question_id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "question_text": {
                                    "type": "string"
                                },
                                "answers": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "answer_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "answer_text": {
                                                "type": "string"
                                            },
                                            "is_correct": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}